v6-bridge/requirements.txt
vantage6-node>=4.5.0
flask>=3.0.0
pyyaml>=6.0
